Use uint8 for network version

+++ internal/blockchain/ethereum/abi_definitions.go
@@ -104,8 +104,8 @@ var networkVersionMethodABI = ABIElementMarshaling{
 	Inputs:          []ABIArgumentMarshaling{},
 	Outputs: []ABIArgumentMarshaling{
 		{
-			InternalType: "uint256",
-			Type:         "uint256",
+			InternalType: "uint8",
+			Type:         "uint8",
 		},
 	},
 }
